Add informal explanation to the README.
I made an electronic voting protocol that is resistant to timing analysis!

Bryan told me to not waste too much time on voting protocols and to focus on
implementing something. So I'm sure he will be thrilled to learn that I ended up
making a voting protocol that I didn't finish implementing!